requirements.txt
flask
sqlalchemy
azure-storage-queue